requirements.txt
# MCP SDK
mcp>=1.0.0

# CrowdStrike SDK
crowdstrike-falconpy>=1.4.0

# Configuration management
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
